requirements.txt
Django==2.0.1
django-cors-headers==2.4.0
django-webpack-loader==0.6.0
gunicorn==19.9.0
httplib2==0.11.3
Pillow==5.2.0
psycopg2==2.7.5
python-instagram==1.3.2
pytz==2018.5
djangorestframework==3.9.0
requests==2.20.0
simplejson==3.16.0
six==1.11.0